+++ Runtime/API/RuntimeGlyphCollection.cs
@@ -2,6 +2,7 @@ using JetBrains.Annotations;
 using Rewired;
 using System;
 using System.Collections.Generic;
+using UnityEngine;
 
 namespace LMirman.RewiredGlyphs
 {
@@ -9,22 +10,25 @@ namespace LMirman.RewiredGlyphs
 	{
 		internal static readonly RuntimeGlyphCollection Default = new RuntimeGlyphCollection();
 
-		/// <summary>
-		/// Automatically generated glyphs for inputs that have no glyph defined anywhere.
-		/// </summary>
-		private readonly Dictionary<string, Glyph> fallbackGlyphs = new Dictionary<string, Glyph>();
+		private readonly Dictionary<int, Glyph> defaultGamepadMap = new Dictionary<int, Glyph>();
+		private readonly Dictionary<int, Glyph> keyboardMap = new Dictionary<int, Glyph>();
+		private readonly Dictionary<int, Glyph> mouseMap = new Dictionary<int, Glyph>();
+
 		/// <summary>
 		/// Glyph mapping based on hardware specific glyphs and hardware specific input ids, found via hardware Guid.
 		/// </summary>
-		private readonly Dictionary<Guid, Dictionary<int, Glyph>> guidGlyphMaps = new Dictionary<Guid, Dictionary<int, Glyph>>();
-		/// <summary>
-		/// Glyph mapping based on hardware specific glyphs and hardware specific input ids, found via hardware type definition.
-		/// </summary>
-		private readonly Dictionary<HardwareDefinition, Dictionary<int, Glyph>> hardwareGlyphMaps = new Dictionary<HardwareDefinition, Dictionary<int, Glyph>>();
+		private readonly Dictionary<Guid, Dictionary<int, Glyph>> joystickGuidMaps = new Dictionary<Guid, Dictionary<int, Glyph>>();
+
 		/// <summary>
 		/// Glyph mapping based on template glyphs and template input ids.
 		/// </summary>
 		private readonly Dictionary<SymbolPreference, Dictionary<int, Glyph>> templateGlyphMaps = new Dictionary<SymbolPreference, Dictionary<int, Glyph>>();
+
+		/// <summary>
+		/// Automatically generated glyphs for inputs that have no glyph defined anywhere.
+		/// </summary>
+		private readonly Dictionary<string, Glyph> fallbackGlyphs = new Dictionary<string, Glyph>();
+
 		internal readonly GlyphCollection collection;
 		/// <inheritdoc cref="InputGlyphs.NullGlyph"/>
 		internal readonly Glyph nullGlyph;
@@ -36,9 +40,6 @@ namespace LMirman.RewiredGlyphs
 		private RuntimeGlyphCollection()
 		{
 			collection = null;
-			guidGlyphMaps.Clear();
-			hardwareGlyphMaps.Clear();
-			templateGlyphMaps.Clear();
 			uninitializedGlyph = new Glyph("Uninitialized", type: Glyph.Type.Uninitialized);
 			unboundGlyph = new Glyph("Unbound", type: Glyph.Type.Unbound);
 			nullGlyph = new Glyph("Null", type: Glyph.Type.Null);
@@ -49,46 +50,61 @@ namespace LMirman.RewiredGlyphs
 			this.collection = collection;
 
 			// Create guid glyph lookup
-			guidGlyphMaps.Clear();
 			foreach (GlyphCollection.GuidEntry guidEntry in collection.GuidMaps)
 			{
-				guidGlyphMaps[guidEntry.GuidValue] = guidEntry.glyphMap.CreateDictionary();
-				foreach (Glyph glyph in guidEntry.glyphMap.Glyphs)
+				if (guidEntry.GlyphMap == null)
 				{
-					// HACK: Big assumption here that every GUID reference is a Joystick.
-					// I am not aware of a way to determine controller type of controllers not plugged in.
-					// Maybe we can load from the Controller data file?
-					glyph.ControllerType = ControllerType.Joystick;
+					Debug.LogError($"No glyph map defined for a \"{guidEntry.ControllerType}\" controller map on collection \"{collection.name}\"", collection);
+					continue;
 				}
-			}
 
-			// Create hardware glyph lookup
-			hardwareGlyphMaps.Clear();
-			foreach (GlyphCollection.HardwareEntry entry in collection.HardwareMaps)
-			{
-				hardwareGlyphMaps[entry.hardwareDefinition] = entry.glyphMap.CreateDictionary();
-				foreach (Glyph glyph in entry.glyphMap.Glyphs)
+				switch (guidEntry.ControllerType)
 				{
-					glyph.ControllerType = entry.hardwareDefinition switch
-					{
-						HardwareDefinition.Unknown => null,
-						HardwareDefinition.Keyboard => ControllerType.Keyboard,
-						HardwareDefinition.Mouse => ControllerType.Mouse,
-						_ => ControllerType.Joystick
-					};
+					case ControllerType.Keyboard:
+						keyboardMap = guidEntry.GlyphMap.CreateDictionary();
+						break;
+					case ControllerType.Mouse:
+						mouseMap = guidEntry.GlyphMap.CreateDictionary();
+						break;
+					case ControllerType.Joystick:
+					case ControllerType.Custom:
+						joystickGuidMaps[guidEntry.GuidValue] = guidEntry.GlyphMap.CreateDictionary();
+						break;
+					default:
+						throw new ArgumentOutOfRangeException();
+				}
+
+				foreach (Glyph glyph in guidEntry.GlyphMap.Glyphs)
+				{
+					glyph.ControllerType = guidEntry.ControllerType;
 				}
 			}
 
 			// Create template glyph lookup
-			templateGlyphMaps.Clear();
 			foreach (GlyphCollection.TemplateEntry entry in collection.TemplateMaps)
 			{
-				templateGlyphMaps[entry.symbolPreference] = entry.glyphMap.CreateDictionary();
-				foreach (Glyph glyph in entry.glyphMap.Glyphs)
+				if (entry.GlyphMap == null)
+				{
+					Debug.LogError($"No glyph map defined for a \"{entry.SymbolPreference}\" template map on collection \"{collection.name}\"", collection);
+					continue;
+				}
+
+				templateGlyphMaps[entry.SymbolPreference] = entry.GlyphMap.CreateDictionary();
+				foreach (Glyph glyph in entry.GlyphMap.Glyphs)
 				{
 					// We can safely assume any template entry is a joystick since template maps are used exclusively by that controller type.
 					glyph.ControllerType = ControllerType.Joystick;
 				}
+			}
+
+			// Create default glyph lookup
+			if (collection.DefaultGamepadTemplateMap != null)
+			{
+				defaultGamepadMap = collection.DefaultGamepadTemplateMap.CreateDictionary();
+			}
+			else
+			{
+				Debug.LogWarning($"No default gamepad template glyph map found on \"{collection.name}\". Some Joystick devices may not be able to show glyphs.");
 			}
 
 			uninitializedGlyph = collection.UninitializedGlyph;
@@ -106,31 +122,34 @@ namespace LMirman.RewiredGlyphs
 		/// Usage of this method is not recommended in most cases and should only be used if you need fine control over glyph display.<br/><br/>
 		/// You are encouraged to use <see cref="InputGlyphs.GetJoystickGlyph(int,Rewired.Controller,Rewired.Pole,out Rewired.AxisRange,int,bool, string)"/> instead.
 		/// </remarks>
-		/// <param name="hardwareGuid">The hardware guid that the <see cref="elementID"/> maps to</param>
-		/// <param name="elementID">The element input id to get a glyph for</param>
+		/// <param name="controllerType">The type of controller to get a glyph for</param>
+		/// <param name="hardwareGuid">The hardware guid for the controller</param>
+		/// <param name="elementID">The element input id to get a glyph for on the <see cref="GlyphMap"/></param>
 		/// <returns>The found <see cref="Glyph"/> inside of this hardware's glyph map. Returns null (<b>NOT</b> <see cref="nullGlyph"/>) if none is found.</returns>
 		[CanBeNull]
-		public Glyph GetNativeGlyphFromGuidMap(Guid hardwareGuid, int elementID)
+		public Glyph GetNativeGlyphFromGuidMap(ControllerType controllerType, Guid hardwareGuid, int elementID)
 		{
-			bool hasGuidGlyphMap = guidGlyphMaps.TryGetValue(hardwareGuid, out Dictionary<int, Glyph> value);
-			return hasGuidGlyphMap && value.TryGetValue(elementID, out Glyph glyph) ? glyph : null;
+			Dictionary<int, Glyph> guidMap = GetGuidMap(controllerType, hardwareGuid);
+			return guidMap != null && guidMap.TryGetValue(elementID, out Glyph glyph) ? glyph : null;
 		}
 
-		/// <summary>
-		/// Retrieve a glyph for this element id that belongs to a specific hardware setup, via hardware type.
-		/// </summary>
-		/// <remarks>
-		/// Usage of this method is not recommended in most cases and should only be used if you need fine control over glyph display.<br/><br/>
-		/// You are encouraged to use <see cref="InputGlyphs.GetJoystickGlyph(int,Rewired.Controller,Rewired.Pole,out Rewired.AxisRange,int,bool, string)"/> or <see cref="InputGlyphs.GetKeyboardMouseGlyph(int,Rewired.Pole,out Rewired.AxisRange,int,bool, string)"/> instead.
-		/// </remarks>
-		/// <param name="controller">The hardware type that the <see cref="elementID"/> maps to</param>
-		/// <param name="elementID">The element input id to get a glyph for</param>
-		/// <returns>The found <see cref="Glyph"/> inside of this hardware's glyph map. Returns null (<b>NOT</b> <see cref="nullGlyph"/>) if none is found.</returns>
+		/// <inheritdoc cref="GetNativeGlyphFromGuidMap(Rewired.ControllerType,System.Guid,int)"/>
 		[CanBeNull]
-		public Glyph GetNativeGlyphFromHardwareMap(HardwareDefinition controller, int elementID)
+		public Glyph GetNativeGlyphFromGuidMap(Controller controller, int elementID)
 		{
-			bool hasHardwareGlyphMap = hardwareGlyphMaps.TryGetValue(controller, out Dictionary<int, Glyph> value);
-			return hasHardwareGlyphMap && value.TryGetValue(elementID, out Glyph glyph) ? glyph : null;
+			return GetNativeGlyphFromGuidMap(controller.type, controller.hardwareTypeGuid, elementID);
+		}
+
+		[CanBeNull]
+		private Dictionary<int, Glyph> GetGuidMap(ControllerType controllerType, Guid hardwareGuid)
+		{
+			return controllerType switch
+			{
+				ControllerType.Keyboard => keyboardMap,
+				ControllerType.Mouse => mouseMap,
+				ControllerType.Joystick or ControllerType.Custom => joystickGuidMaps.GetValueOrDefault(hardwareGuid),
+				_ => throw new ArgumentOutOfRangeException(nameof(controllerType), controllerType, null)
+			};
 		}
 
 		/// <summary>
@@ -147,7 +166,7 @@ namespace LMirman.RewiredGlyphs
 		public Glyph GetNativeGlyphFromTemplateMap(SymbolPreference symbolPreference, int templateElementID)
 		{
 			bool hasTemplateGlyphMap = templateGlyphMaps.TryGetValue(symbolPreference, out Dictionary<int, Glyph> templateGlyphMap);
-			return hasTemplateGlyphMap && templateGlyphMap.TryGetValue(templateElementID, out Glyph glyph) ? glyph : null;
+			return hasTemplateGlyphMap && templateGlyphMap.TryGetValue(templateElementID, out Glyph glyph) ? glyph : defaultGamepadMap.TryGetValue(templateElementID, out glyph) ? glyph : null;
 		}
 
 		/// <summary>